refactor: 외부 라이브러리를 도메인에서 분리
