+++ src/main/java/subway/business/domain/PathCalculator.java
@@ -1,0 +1,9 @@
+package subway.business.domain;
+
+import java.util.List;
+
+public interface PathCalculator {
+    int getTotalDistance(Station sourceStation, Station destStation);
+
+    List<Station> getShortestPath(Station sourceStation, Station destStation);
+}
